requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
python-dotenv==0.21.0
requests==2.28.1
requests-file==1.5.1
six==1.16.0
urllib3==1.26.13
yarl==1.8.2
